Add abilities back to Species
f547aa4 removed this instead of just making the set's ability
override the default. A future commit will be making ability in the
Pokemon constructor optional and needs this information present in
order to be able to fill in a default.

This data was generated by programmatically going through PS's
Ability data for the species and choosing the ability with the
highest rating + some manual adjustments where necessary.

Verified #276 is still fixed after this change.