Introduce set/weight import logic (#346)
Adds logic for pulling sets and 'weights' from Smogon, as well as for pulling sets from other third party sets. Currently supports more formats than the calc does, however later discards the formats it doesn't understand when running in `legacy` mode (the default).

Going forward, the sets format will be keyed first by format in order to facilitate the dropdown logic discussed on #343. The JSON sets will also be pretty printed with whitespace in the repo because Parcel will handle minimizing it for us.

'Weights' are intended to be used to modify the sort order of dropdowns, coupled with information already existing in sets (options that appear in sets will be sorted first by usage, then the remaining options will be sorted by global usage for the format). 

The commit fixes the typo of 'Heracrossite' and adds 'Crucibellite', but `missing.json` contains all the IDs we're currently missing (needed for name lookup given usage stats currently report ids in most places in the `chaos` reports). Once we close #334, `missing.json` and the logic around it can be deleted, and `sets/missing > sets/src/missing.json` can be run to to update as the new effects are supported.

Fixes #343