Cleanup damage description handling
The damage type is pretty obtuse to use (though critical to have to
be able to fully implement everything), but most users will just
care about the highest and lowest values so this provides a
Result#range() method for them. This also means
expect(...).toBeRange() can be removed which is a nice bonus, as
well as the duplicated minMaxDamage method in the UI and package.

Furthermore, this change takes care to handle overflow cases where
the description can be misleading - instead by sorting (or taking
min, max in one location for speed) we can ensure the descriptions
always make sense. We don't *always* sort because overflow is super
rare and theres no point in doing an O(nlogn) operation all the
time, and furthermore we take care to not mutate the raw damage
(we use slice() to copy) because we want to display the rolls in the
order they appear.