Update sets
This is not exactly a clean update. Due to my not-so-good internet
connection I'm not able to run the full importer and chaos' setup seems
to not be able to generate gen9 sets. So, I ran the importer script
on only gen9 and disabled Smogon sets (only stats), then I added any new
Pokemon in those sets to the gen8 sets.