Update items, relax test on move data
Energy Powder removed, this doesn't seem to be in PS' data.

In the tests, move data is compared to a different source, but this source can
have extraneous `<property>: undefined` fields that do not match against the
actually missing fields in the calc data.

This also lets Hyper Drill have breaksProtect: true, I'm not sure about the
technicalities of this flag but I didn't want to break the calc if it isn't
concerned with the distinction of "bypasses but does not break Protect". Maybe
some other moves are also allowed with technically incorrect additional
flags. Someone will have to check.