Rewrite the set importer
The new set importer uses `data.pkmn.cc` as the data source which
handles a lot of quirks for us (such as converting the Smogon formats
to PS formats, having the data nicely laid out).

The new set importer is a lot faster too and won't consume GBs of
internet due to not downloading every analysis page.

These bugs should be fixed (and more that I won't list out):
- Zacian-C and Zamazanta-C not importing
- SV BH/ND sets not importing