Import: Prettify set output
We would like to have meaningfull diffs for set update commits but
`JSON.stringify` can only output either in one line or with a newline
for every single element which occupies *way* too much space. So, we
have to stringify the calc sets on our own so it produces useful diffs
in a reasonable amount of lines

The output looks like:

```
  "specie": {
    "set": {...set info},
    "set2": {...set info}
  }
```

This output allows us to easily see what sets were removed/added when
debugging or just making sure nothing is wrong.

The build script will minify the sets now so this won't have any
impact on users of the calc