Remove misleading KO chance outputs (#536)
This removes a good amount of the text duplication in getKOChance() , but really the point is to resolve #448: where previously 100% chance to be KOed would be returned will instead yield 99.9%, and 0% will instead show 0.1%.

I also added a test for the specific case mentioned. I can't think of an example for 0%, but I have seen it before.