Support building from shared template
This makes it much easier to update the header and footer without
needing to edit every file individually.

Since the src template files seem to be designed to be opened in-browser
without a build step, I've done my best to replicate that, which means
that changes will still need to be committed across all four files.
But at least edits will be easier.